refactor: HttpResponseUtils를 생성하여 response generate 분리
